Change everything
I found some code online that allows me to generate an array of all
characters available on a keyboard. I did this to save time manually
inputing each character. I also converted that array into a string and
created a for loop that pulls randomly characters from that string a
certain number of times. This was an assembly of code on stack overflow.